+++ tests/bdd/main.rs
@@ -1,5 +1,6 @@
 mod auth_query_helper;
 mod binary_upgrade_fd_helper;
+mod blackhole_helper;
 mod doorman_helper;
 mod extended;
 mod fallback_helper;